Refactor - move league statistics methods and tests into GAME MANAGER; updated stattraker
