+++ test/league_statistics_test.rb
@@ -2,43 +2,39 @@ require "./test/test_helper"
 
 
 class LeagueStatisticsTest < Minitest::Test
-  def setup
-    @data = { games: './data/games.csv',
-              teams: './data/teams.csv',
-              game_teams: './data/game_teams.csv'
-            }
-    @league_stats = LeagueStatistics.new(@data)
-  end
-
-  def test_it_exists
-    assert_instance_of LeagueStatistics, @league_stats
-  end
-
-  def test_number_of_teams
-    assert_equal 32, @league_stats.number_of_teams
-  end
-
-  def test_best_attackers
-    assert_equal 54, @league_stats.best_attackers
-  end
-
-  def test_worst_attackers
-    assert_equal 7, @league_stats.worst_attackers
-  end
-
-  def test_most_goals_by_away_team
-    assert_equal 6, @league_stats.most_goals_by_away_team
-  end
-
-  def test_most_home_goals_by_team
-    assert_equal 54, @league_stats.most_home_goals_by_team
-  end
-
-  def test_least_visitor_goals_by_team
-    assert_equal 27, @league_stats.least_visitor_goals_by_team
-  end
-
-  def test_least_home_goals_by_team
-    assert_equal 7, @league_stats.least_home_goals_by_team
-  end
+  # def setup
+  #   @data = { games: './data/games.csv',
+  #             teams: './data/teams.csv',
+  #             game_teams: './data/game_teams.csv'
+  #           }
+  #   @league_stats = LeagueStatistics.new(@data)
+  # end
+  #
+  # def test_it_exists
+  #   assert_instance_of LeagueStatistics, @league_stats
+  # end
+  #
+  # def test_best_attackers
+  #   assert_equal 54, @league_stats.best_attackers
+  # end
+  #
+  # def test_worst_attackers
+  #   assert_equal 7, @league_stats.worst_attackers
+  # end
+  #
+  # def test_most_goals_by_away_team
+  #   assert_equal 6, @league_stats.most_goals_by_away_team
+  # end
+  #
+  # def test_most_home_goals_by_team
+  #   assert_equal 54, @league_stats.most_home_goals_by_team
+  # end
+  #
+  # def test_least_visitor_goals_by_team
+  #   assert_equal 27, @league_stats.least_visitor_goals_by_team
+  # end
+  #
+  # def test_least_home_goals_by_team
+  #   assert_equal 7, @league_stats.least_home_goals_by_team
+  # end
 end
